runJob and help functions now in their own files
